Moving logic into respective modules and hiding accessors
